[FIX] knowledge: rename outline command
This commit renames the /outline command inside knowledge from Outline to
Article Structure.
This is done as such because the translation in FR of Outline used the same
translation as for the tables: *Contour*, instead of *Structure*.

Thus this commit changes the name of the command and modifies the .pot file
for the traductions.

task-3169027

closes odoo/enterprise#36753

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>