[FIX] documents_project: hide document smart button on subtask creation
Current behaviour:
If you click on the Document smart button during a subtask creation,
you are met with a traceback.

Expected behaviour:
After confirmation with the PO, we should hide the smart button,
since the task is not yet created.

Steps to reproduce:
- Install Project and Documents
- Activate Subtasks in Settings
- On an task: Subtasks > Add a line > Create/New > Docs smart button

Reason for the problem:
Since the task is not yet created, there is no id yet for the task,
so when triggering the action to open the document triggers an error
because the `active_id` doesn't exist.

Fix:
Set the button invisible if the `id` of the task is False (aka not
yet created).

Affected versions:
- 16.0
- saas-16.1
- master

opw-3133040

closes odoo/enterprise#36759

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>