[IMP] account_report: Adding title, vat number, options and address
Before this PR, when printing a report the title was missing, this has already been addressed in this PR: https://github.com/odoo/enterprise/pull/36163.
In addition, we had to add the vat number and the address in the header of the report. This PR also add the options of the reports in the pdf.

Task-id: 3171683
Part-of: odoo/enterprise#36760