[IMP] delivery_*: unify label/doc prefixes
Prior to this commit the shipping label/doc attachment names were
inconsistent and difficult to distinguish. To avoid this, now we enforce
consistent prefixing of the filenames that is now reuseable across
shipping connectors (i.e. we copy the return label prefix method).

Now all label documents will be prefixed with "Label" to make them
obvious and all other shipping provided documents will be prefixed with
"ShippingDoc-". To distinguish between the type of label, "Shipping" and
"Return" are included in their prefixes as well.

Additionally, improve the English of the chatter message created when
the shipment is created (i.e. when the shipping labels are created)
since "into" was grammatically incorrect.

Part of task: 3046178
COM PR: odoo/odoo#126791

Part-of: odoo/enterprise#43362