[FIX] helpdesk_mail_plugin: create a record with the partner's company
Before this commit: if the partner belonged to "Company B" and the
default company was "Company A", it would cause an issue during
the creation of the record (opportunity, ticket, ...) from the mail plugin.

Community PR: https://github.com/odoo/odoo/pull/112088

The solution is to use the partner's company to create.

opw-3141578

closes odoo/enterprise#39870

X-original-commit: f47c94ec31f8bb724b77931d63ac46fb3d067080
Related: odoo/odoo#118771
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>