[FIX] l10n_lu_reports: fix balance sheet report incorrect calculation
Steps to reproduce:

- Install Accounting and the l10n_lu module.
- Go to Accounting > Reporting > Statement Reports > Balance Sheet.
- In the top bar of the report, select Balance Sheet (abridged)(LU) as
the report template.
- Now scroll down until "a) becoming due and payable within one year"
and open the drop-down.

Issue:

We can see that the value we have inside does not match what we are
showing we are taking into account the Customers invoice twice.

Solution:

Fixed the formula that calculates this part to replace what's causing
the issue for the proper one.

opw-3150153

closes odoo/enterprise#36823

X-original-commit: e67c889996449a0170823d707a2a65e73f30bd56
Signed-off-by: Nicolas Viseur (vin) <vin@odoo.com>