[FIX] knowledge: hide caret in sidebar if all children are items
Purpose:
========
Currently, if a child article has only items as child articles, the caret
for this article is shown after unfolding its parent.

This commit hides the caret in this case, as there are no articles to unfold
if the children articles are all items (since items are not shown in the
sidebar).

Task-3168666

closes odoo/enterprise#36779

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>