[FIX] hr_contract_salary_payroll, l10n_be_hr_contract_salary: copy needed fields on simulation link
When generating a simulation link a couple of fields
(time_credit_type_id and standard_contract_id) weren't correctly copied
from the previous contract. This commit fixes that.

Also, in the case of working schedule change, the standard_calendar_id
was incorrectly set as the previous contract's calendar, instead of the
'full time resource calendar', which is wrong in the case of two
consecutive working schedule changes.

task-2985141

closes odoo/enterprise#39149

X-original-commit: 624f137ae77cfc12ff23e0c1a82bc8d2ee1f6d96
Signed-off-by: Kevin Baptiste <kba@odoo.com>