[FIX] hr_contract_salary: convert final_yearly_costs to computed field
In the generate_simulation_link wizard, the final_yearly_costs field was
computed through an onchange on contract_id. However, this did not seem
to be trigger without manually resetting contract_id. With this change
this field is converted into a computed field (which doesn't have this
issue) that is non-read-only, so it can be manually overridden.

task-3177146

closes odoo/enterprise#36794

Signed-off-by: Kevin Baptiste <kba@odoo.com>