[FIX] documents : url redirection video documents
In the shared space, shared videos linked to an url (i.e. youtube link)
were treated as mp4 videos that were downloadable. The user was then
trying to download an url, which was not working. After the fix, the
behaviour expected is the redirection to the corresponding url when
trying to preview. The 'download' button has been replaced by a 'link'
button.

TaskId-3167523

closes odoo/enterprise#36798

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>