[FIX] mrp_workorder: use another SN than the reserved one on a WO
On a MO, if a WO consumes several times the same tracked-by-sn
component, and if the user consumes the SN in a different order than
the default one, it will create an inconsistency on the SNs' reservation

To reproduce the issue:
1. Create two storable products P_compo, P_finished
   - P_compo:
     - Tracking: Serial
2. Update on-hand qty of P_compo:
   - 1 x SN01
   - 1 x SN02
3. Create a BoM:
   - Product: P_finished
   - Operations:
     - A new operation OP
   - Components:
     - 2 x P_compo, consumed in OP
4. Confirm a MO for 1 x P_finished
5. Open the tablet view of OP
   - SN01 should be the suggested
6. Set SN02
7. Continue consumption
   - Error01: SN02 is suggested but it has been used in the step
     before. Also, if the user opens the detailed operations of the
     raw move, he will see that the reservation of SN01 does not
     exist anymore
8. Set SN01
9. Validate
10. Check the detailed operations of the raw move

Error: The situation is not that consistent because there is now a
reservation for SN01 but not for SN02

Once the MO is confirmed, both SN are reserved. Step 6, we assign
SN02 on the first SML (i.e. the one that reserves SN01). As a result,
we unreserve SN01 and try to reserve SN02 (this reservation will not
work since SN02 is already reserved by the second SML)

Then, step 8: we change the lot of the second SML so we first
unreserve it (-> we unreserve SN02) then we try to reserve the new
SN (-> it reserves SN01)

When we assign a lot on a SML, we should check if there already
exists another SML that reserves the lot.

OPW-3104930

closes odoo/enterprise#36856

X-original-commit: 87167334f331da6260ebf060505b9c5ed4f1d293
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Adrien Widart <awt@odoo.com>