[IMP] account_accountant: Improve perf of bank reco widget
When mounting a batch payment of 1000 lines, it takes a while (+- 12s).
Suppose we are looping through the 1000 lines and call the field 'tax_tag_ids' computed by:

    @api.depends('source_aml_id')
    def _compute_tax_tag_ids(self):
        for line in self:
            if line.flag == 'aml':
                line.tax_tag_ids = [Command.set(line.source_aml_id.tax_tag_ids.ids)]

The problem is this compute is not assigning the value everytime.
Suppose all lines have a flag differents than 'aml'.
For the first line, the orm will call the compute for the whole recordset (1000 records) that will assign nothing.
For the second line, the value is still not computed so the orm will call again this compute and so, 1000 times.
So instead of calling once the compute for 1000 lines, the compute is called 1000 times with the whole recordset.

By forcing a value everytime, the time is reduced down to < 3s.

X-original-commit: fe297d7b8b39d0b7945b453bcb58b70fe8e3f6c8
Part-of: odoo/enterprise#36999