[FIX] appointment: wysiwyg: don't close full composer...
when inserting an appointment link.

This commit fixes an issue introduced by [1]: in a form view with
chatter (e.g. crm.lead), click on Send message, open the full
composer, in the dialog, type / in the message body and select
Appointment in the dropdown (it should open a dialog), set a value
in the record and click on Insert Link. Before this commit, the
two dialogs were closed, so basically, it wasn't possible to add
an appointment link in a message from the full composer.

The reason it closes the full composer dialog is because the clicked
button is 'special="save"', which means that the record will be
saved, and an "ir.actions.act_window_close" will be executed, which
closes the dialog.

This commit fixes the issue by bypassing the normal framework
behavior for the two footer buttons of the appointment dialog, s.t.
the record is saved and the link is added (if necessary), and the
appointment dialog (and only that one) is closed.

[1] 121a7717e3e0b104ac6726681f1f009774e4f8c9

Part of opw 3144747

closes odoo/enterprise#36878

X-original-commit: 070da2e3d6523ace8977328f7183922ce8da902b
Signed-off-by: Georis François (fge) <fge@odoo.com>