[FIX] knowledge: allow to add readonly articles to favorites
Purpose:
========
Previously, clicking on the toggle favorite button while on a readonly
article threw an access error.
This happened because we tried to update the record while not having write
access on it.
Now, the toggle button uses the `action_toggle_favorite` method of the
knowledge article model instead.

Task-3175460

X-original-commit: 89461540ad23c5bd3ba0e71a0ed603177fb6d86b
Part-of: odoo/enterprise#37437