[FIX] sale_subscription: prevent double mail while auto closing
Before this commit, when subscriptions had to be closed because the
customer did not paid, a mail was sent before the closing.

As the email are sent after the sql commit, if something went wrong during the
commit, the mails would be sent twice if the cron was restarted.
The same contract would be selected for auto closing and the same mail
would be sent. This commit send the mail before the commit when we close the
contracts in batches. if something goes wrong, it is reverted and no mail is sent.

Part-of: odoo/enterprise#36834