[FIX] account_accountant: Fix auto_balance with no account on partner
When an auto_balance line is created on the bank reconciliation widget, if the partner is specified, we create an open balance using the receivable/payable account of the partner instead of the suspense account.
In case the accounts are not set on the partner, an open balance was created with an empty account.
This commit improves that by creating a regular suspense line in that case.

X-original-commit: e69b7e7886393f337557a9ccadcf3c7105ced612
Part-of: odoo/enterprise#36999