[FIX] web_enterprise: WebManifest & ServiceWorker route auth
Once the user access the backend (aka. `/web`), the browser loads the
WebManifest and the ServieWorker is registered. After a logout, he's
redirected to the login page (aka. `/web/login`).

But this path being still under the ServiceWorker's scope (cf. `/web`),
the browser continues to try to check for an updated ServiceWorker.

If the user (same or another one) tries to log back in, a race condition
appears between the login request and the check for the ServiceWorker.
This issue is due to the ServiceWorker's route requiring an
authenticated user, which is not (yet) the case in this scenario,
conflicting with the login request.

This commit fixes it by reverting the `auth="public"` of the
ServiceWorker route. A similar issue, even if less likely to occur,
affects the WebManifest route for which the same change is applied.

Regarding the WebManifest route, shortcut's collection still requires an
authenticated user, but we fallback to an empty list if not authorized
(not authenticated, no access to apps, not enough rights...).

Note: the ServiceWorker, once registered, is entirely managed by the
browser ; we don't have the ability to control when the browser will
attempt to update it.

> Browsers will check for updates to a service worker when:

> * The user navigates to a page within the service worker's scope.
> * navigator.serviceWorker.register() is called with a URL different from
>   the currently installed service worker—but don't change a service
>   worker's URL!
> * navigator.serviceWorker.register() is called with the same
>   URL as the installed service worker, but with a different scope. Again,
>   avoid this by keeping the scope at the root of an origin if possible.
> * When events such as 'push' or 'sync' have been triggered within the
>   last 24 hours—but don't worry about these events yet.

Steps to reproduce:
- Login to an Odoo instance
- Logout
- Try to go to ´/web/login´ and login again.

References:
https://developer.chrome.com/docs/workbox/service-worker-lifecycle/#handling-service-worker-updates
https://web.dev/manifest-updates/

closes odoo/enterprise#36830

Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>