[FIX] timesheet_grid: add default next date
Issue:
When we install the timesheet module,
add fields to models `res.company` and `res.config.settings`.

The fields that are used to trigger the sending of mail for timesheets reminder (via cron)
are `timesheet_mail_employee_nextdate` and `timesheet_mail_manager_nextdate`.
These fields need to be updated via a calculation (because they depend on a delay and an interval).

Without the fix, these fields won't have any value
as long as the user does not modify the record by changing the delay or interval (via `res.config.settings`)
in the Timesheets application settings (which trigger the `write` function and thus the `_timesheet_postprocess` function).

Therefore, if the user does not make any changes to the settings, no mail will be sent.

Solution:
Give a default value (via a post init function)
to the `timesheet_mail_employee_nextdate` and `timesheet_mail_manager_nextdate` fields
when installing the `timesheet_grid` module.

opw-3114331

closes odoo/enterprise#36938

X-original-commit: fced33d8768fdda8c3dded98465b77b9d8b489ae
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>