[FIX] website_sale_renting: can add to cart rental products
Current behaviour:
A visitor on the ecommerce cannot add a rental product to the cart,
if said product is only checked with "can be rented", but not
"can be sold".

Expected behaviour:
After PO's confirmation, visitor should be able to add to cart
rental products that are only set as "can be rented".

Steps to reproduce:
- Install Ecommerce, Rental, Sales
- Choose a rental product, set it as "can be rented" and unset "can
  be sold", publish it to the website.
- Logout and try to add the product on the eCommerce, it will fail
  to add the product. (With a warning that the product doesn't exist,
  this is why it cannot be added to cart)

Reason for the problem:
The method that checks if a product can be added to the cart (`_is_add_to_cart_allowed`)
is requiring the "can be sold" checks to be `true` for it to be added.

Fix:
Override `_is_add_to_cart_allowed` and add a disjunction with the "can be rented"
flag.

Affected versions:
- 16.0
- saas-16.1
- master

opw-3160771

closes odoo/enterprise#36851

Signed-off-by: William Braeckman (wbr) <wbr@odoo.com>