[IMP] l10n_mx_edi_pos: add extra mexican fields to pos flow
In the POS, allow to fill the `l10n_mx_edi_fiscal_regime` and the
`l10n_mx_edi_no_tax_breakdown` fields on the `PartnerDetailsEdit`
window.

In addition, when creating an invoice from a POS order (for the current
order or for a previous order when the customer ask for an invoice),
display a popup to fill the `l10n_mx_edi_usage` and the
`l10n_mx_edi_cfdi_to_public` fields.

task-3178954

closes odoo/enterprise#36859

Related: odoo/odoo#112285
Signed-off-by: Laurent Smet <las@odoo.com>