[FIX] stock_barcode{,_mrp}: explode kit-SM on validation
[1] and [2] have been deployed so a user can add scan a kit in the
barcode app. However, under some conditions, a traceback appeared.
So, we revert [1] (see oc-side PR linked to this commit) and change
the fix for a new one: if needed, we send RPC to explode the kit-SM.

In the test, the commit adds a step: click on the pencil to edit the
line. This was the origin of the traceback: without the new fix
version, if the picking is immediate, clicking on the pencil will
trigger a save and this save will trigger the explode. As a result, we
try to edit a line that does not exist anymore (because of the
explode), hence the traceback.

[1] odoo/odoo@879ef5ef6bc06e6db46f3e5f9285619316f183e8
[2] 2bdbf690acf31355ccb8c71190e4e2de0eb49f66

OPW-3015933

closes odoo/enterprise#36964

X-original-commit: 47c3f8b000fe4afa664b63d73e484de86fe17a48
Related: odoo/odoo#112428
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Adrien Widart <awt@odoo.com>