[FIX] l10n_mx_edi: fix and improve performance on edi_origin search
Before this commit:

The search domain was using the `like` operator which is a bad idea
in this case as we also use the `%` symbol.

e.g: `[('l10n_mx_edi_origin', 'like', '04|%')]`
Would be transformed in SQL to `... WHERE l10n_mx_edi_origin like '%04|%%' ...`
which is not what we want as it was intended to be uses as a `.startswith`.
In practice the result will the same in this case as the pipe separator is not
used elsewhere, but the performance decrease.
In this case we should use `[('l10n_mx_edi_origin', '=like', '04|%')]` instead.

After the commit:

Performance analysis (made on 15 as client database ion this version):
 On the customer instance with ~3250 records (`account_move where l10n_mx_edi_origin is not null`), to evaluate `_compute_l10n_mx_edi_cancel` of one record:
 - Current code: ~1.4 sec
 - Changing the 2 `like` into `=like`: ~1.3 sec
 - Using 1 `=like`: ~1.4 sec
 - Using `order='id desc'`: ~0.1 sec
We can see that the `order` here is crucial as it did allow to make it 10x faster!
the change to `=like` and the "merge of both `like`" didn't improve the query time.
But was kept as the domain is more readable in that way

opw-3164620

closes odoo/enterprise#36995

X-original-commit: 940aca56bf48c7593f3ca5a28d71915531779709
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Sens Loan (lse) <lse@odoo.com>