[FIX] helpdesk: access error when creating a new team
Previous to this commit, there was an access error
raised when trying to create a team when all the other
teams were archived.

Unfortunately, it was not possible to have the rule
in place and thus it was removed. In order to not have
stages of multiple companies shown to the user, the
stages menu is only shown in debug mode.

These changes also make the helpdesk app more in line
with the project app, which handles task stages
similarly.

Task-3177229

closes odoo/enterprise#37109

X-original-commit: 43ed6c56ab4f3c9dcdcbbf18441d808a5f5574c6
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>