[FIX] appointment: attribute error autovacuum appointment invite
An AttributeError 'calendar.event' object has no attribute 'end' traceback that
occurs in appointment/appointment_invite:_gc_appointment_invite was caught by
Sentry. Because there is no "end" field in "calendar.event" model.

So we use the 'stop' field instead of 'end' because it store stop date and time
of the event

sentry-3891597633

closes odoo/enterprise#36877

Signed-off-by: Achraf <abz@odoo.com>