[FIX] sale_subscription: not reset upsell discount on confirmation
Current behaviour:
When setting the discount on an upsell manually, and then confirming
the upsell, the discount value was reset at the value it was when
creating the upsell.

Expected behaviour:
Should be able to set the discount on the upsell manually and it
should be kept that way, not being reset.

Steps to reproduce:
- Install Subscriptions
- Create a subscription, confirm it, invoice it, upsell it
- Take note on the discount value
- Change the discount manually, save it (should stay changed to what
  you set)
- Confirm the upsell, notice that it reverted back to the value
  we've taken note on.

Reason for the problem:
Since the discount is a precomputed field, when we confirm the upsell,
a new `next_invoice_date` is computed, which triggers the compute of the
discount, which overwrite the manually set value.

Fix:
We remove the dependency on `order_id.next_invoice_date` on `_compute_discount`,
to prevent the computation when we confirm the upsell.

Affected versions:
- 16.0
- saas-16.1
- master

opw-3068419

closes odoo/enterprise#37132

X-original-commit: c3bc690fa9e3c0a852baa9daed6798aa18980f45
Signed-off-by: Arnaud Joset <arj@odoo.com>
Signed-off-by: Piryns Victor (pivi) <pivi@odoo.com>