[FIX] delivery_iot: Avoid duplicated actions execution
Before this commit:
Assuming we have a "Shipping Labels Printer" set on an operation type.
Assuming we also have a delivery carrier which send Label through the chatter of this picking type (like DHL or BPost).
If we have several sessions connected with the same user (author of the chatter message), for examples, on different devices.
Each session will send the same IoT request to print the label (as the code rely on the bus).

In consequences, the same label will be printed multiple times

After this commit:
The label is print only once

Associated community PR: https://github.com/odoo/odoo/pull/112355

opw-3081423

closes odoo/enterprise#39087

X-original-commit: bf386421c1bf2d05f036511d255bf284631309c5
Related: odoo/odoo#117251
Signed-off-by: Loan (LSE) <lse@odoo.com>
Signed-off-by: Sens Loan (lse) <lse@odoo.com>