[FIX] account_avatax: don't send posted moves
To reproduce:
1/ Post an invoice with the Avatax fiscal position
2/ Click “Generate a Payment Link” in the action dropdown
3/ An Avatax error is shown mentioning “DocStatus is invalid for this
operation”

Odoo calls `button_update_avatax` with `commit=False` in the
`payment.link.wizard` _inherit. This shouldn't happen when the invoice
is committed because:

- From Avatax's view we try to "uncommit" because of `commit: False`, and
- We don't want to recompute the taxes anyway, the invoice is posted

Excluding posted invoices in `payment.link.wizard` was considered but is
trickier because it's used both on `sale.order` and `account.move`.
Changing the signature of private functions is allowed but discouraged
in the stable policy. Not changing the signature was also considered but
would be less clean and it's unlikely anyone overrode `_send_to_avatax`.

opw-3136249

closes odoo/enterprise#37704

X-original-commit: 49c282444aa2de77630327105c3dcd04f0e2c116
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: jorenvo <jov@odoo.com>