[FIX] web_studio: disable address form edition
An error is thrown when trying to edit fields in the address of a
partner when an `address_view_id` is set on the company country

Steps to reproduce:
1. Install Studio and Peru - Accounting (l10n_pe)
2. Open any contact in the Contacts app
3. Toggle Studio
4. Click on any field in the address
5. An error is thrown

Solution:
Render the original view (without the `address_view_id` modification)
by using the `no_adress_format` context key

Problem:
The view is edited with the `address_view_id` set on the company
country. Thus we cannot find the fields of pe_partner_address_form
in the original view when we try to edit them.

opw-3136618

closes odoo/enterprise#36928

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>