[FIX] sale_subscription: do not send end_date when closing subscription
Before this commit, when the contract was closed, the end_date was set to today.

It was not an issue but if the contract was reopened it was difficult to know
if the end_date had to me modified/removed.

We decide to not update the end_date when the contract is closed. end_date is
an informational value and it is possible to read the chatter/sale.order.log
to know exactely when the contract was closed.

closes odoo/enterprise#36834

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>