[FIX] project_enterprise: no contract access error when scheduling task
Current behaviour:
If a user that doesn't have any access to contracts, adds a task in
the gantt view and assigns an employee on it, he gets an Access
Error for trying to access hr.contract, but the task is created
nonetheless.

Expected behaviour:
Should not get an Access Error, because the task gets added anyway,
and you can add the task from the list view without any problem.
(the issue is only on the gantt view).

Steps to reproduce:
- Install hr_timesheet, contracts (not attendance)
- On a user, remove all access rights that are HR related, timesheet
  rights should be `access to all employee timesheets` (not admin)
- Create a new project as admin
- Login as the other user
- Go to your new project, gantt view, magnifying glass, create/new,
  assign another user on the task, save it
- You get an access error for the model `hr.contract`.

Reason for the problem:
When computing the schedule for the task, we need to evaluate the
working calendar, which internally makes a few searches on
`hr.contract`, which the current user doesn't have access to.

Fix:
Add a few strategic `sudo()` in `project.task` to ignore the access
requirements for the contracts. This way the client is not be forced
to grant contract access to a user which schedules task and assigns to
other employees on the gantt view.

Affected versions:
- 16.0
- saas-16.1
- master

opw-3142948

closes odoo/enterprise#36940

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>