[IMP] sale_subscription_dashboard : Batch KPIs for subscription report
This PR change the implementation of the sale_subscription_dashboard KPIs view to batch
the retrieval of information.

Before this PR, each point coordinate of the several KPI dashboard graph
were computed with a sql request.
As each graph would draw a point per day, the amount of SQL request to
display the dashboard was overly high (~30 per graphs).

![image](https://user-images.githubusercontent.com/50832879/217816677-f58d178f-3bea-4abf-ac98-78c91087fd99.png)

Production database with a lot of account move lines and subscription
could not be displayed without a time out error.
The KPI dashboard is displaying 9 graphs and the amount of SQL requests
is now approximately divided by 30.

Moreover, the PR allows the function to work with a defined set of dates
instead of a start_date and end_date.

Furthermore, some indexes are added on account_move_lines subscription_start_date
and subscription_end_date and company_id to speed up graph generation.

As a result, the PR provides significant speed-up for the KPI view, especially for
large database.
The specification of the KPI API has also been updated to reflect these changes and
now return an array tuple containing the date and value of each data point.

task-id : 3061203

closes odoo/enterprise#37006

Forward-port-of: odoo/enterprise#35792
X-original-commit: e4f33bd3af761a577e7834cddbe466ed8294d968
Signed-off-by: Arnaud Joset <arj@odoo.com>
Signed-off-by: Damhaut Florian (flda) <flda@odoo.com>