[FIX] knowledge: rewords trash related actions
This commit changes the wording for all the actions linked to the trash as
the old one was confusing for users.

We reworded "Remove from trash" to "Restore from Trash" in the burger menu
and "Remove from Trash" to "Restore" in the banner.

We also changed the pot file to include the translations of the reworded
sentences.

task-3180742

closes odoo/enterprise#36948

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>