[IMP] sale_subscription: Settings & Templates overhaul
Since the merge of Subscriptions with Sales orders,
the concept of "quotation template" has been distorted.
It's no longer used only to pre-fill the SO, one information
required in the life cycle of the subscription is saved in the
template and only in the template:  "closable by customer".
There is also a problem when no template is defined.

Information is hidden for confirmed SO because the
template reference is not visible anymore.

Solutions :
We create a section for subscription settings that contains the following settings :
- Closable by customer
- Start date
- First contract date

In addition, multiple company dependant settings will by available to pre-fill SO
properties when created without templates.
Those settings will also be used as default values when creating
new quotation template.

task-id : 2877341

community : odoo/odoo#112433

closes odoo/enterprise#36968

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>