[FIX] account_reports: fixing failing test
The aim of this commit is to make this test more robust and avoid it
failing for reasons unrelated to this test purpose. (ie reports
mechanics)

See OCO's message on the PR for a detailed explanation of why these
tests used to pass while they shouldn't have.

Original commit e352a0c by @Levizar.

closes odoo/enterprise#36809

Signed-off-by: Arthur Gossuin (goa) <goa@odoo.com>