[REM] test_web_grid: remove module due to the OWL conversion
Before the OWL conversion of the grid view, this module was there to
test all the python methods used only in the grid view as `read_grid`
for instance.

This commit removes the test module for web_grid because the grid will
now used the basic ORM methods as the others views, that is `read_group`
method. Also, the columns are no longer formatted in the python code.
All things used in the UI (except the data fetched) are now built in the
JS code as it is the case for the gantt view for instance.