[IMP] mrp_plm: ECO compare lines w/ same product
Changes the `mrp.eco` method `_get_difference_bom_lines` so it can
compare the old BoM with its revision correclty even when there is
multiple BoM's lines for the same product( + variant)( + operation).

That way, if a BoM has as BoM's lines, for example:
- Product A x1
- Product B x1
- Product A x1

The second line for Product A is no more considered as an additional
line as soon as the revision is created, but will be correctly compared
to the corresponding revision's BoM's line.

task-3089368
opw-2823253

closes odoo/enterprise#35969

Related: odoo/upgrade#4428
Related: odoo/odoo#110227
Signed-off-by: Steve Van Essche <svs@odoo.com>