[FIX] data_merge: 'NoneType' object is not iterable in cron cleanup
If applied, this commit will solve the cron cleanup issue.
The 'NoneType' object is not iterable while evaluating 'model._cron_cleanup()'
this issue occurs in model 'data_merge.record' because of method
_original_records() returns 'None' and _cleanup() method tried to iterate
records. The None object can't be allowed in iteration.

so, we put conditions before using the _original_records() method.

sentry - 3923640623

closes odoo/enterprise#40043

X-original-commit: 426b57d1fb67377d4feecb122e4c0216bb311c50
Signed-off-by: Kevin Baptiste <kba@odoo.com>