[FIX] worksheet: fix access read on compute
When creating a new task as a user with lower privileged and selecting
the "Search More..." on the "Worksheet Template" field will output a
read access error.

How to reproduce
================

1. As admin, create more than 7 worksheets
2. As demo, open the form to create a new task, on the "Worksheet
   Template" field click on "Search More..."

Actual Behavior
===============

The demo user doesn't have the read right on the `model_id` field, this
make the compute of `worksheet_count` automaticly fail.

Expected Behavior
=================

Giving sudo access on the compute should solves the issue and allows an
unprivileged user to access the "Search More..." without any problem.

opw-3163790

closes odoo/enterprise#37329

X-original-commit: 5452036363a8c6feee580fc2767116bb84bf2126
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>