[FIX] account_reports: flush parent state when closing
The generic tax report handler makes use of an sql query to gather the
lines necessary to construct the closing entry. It is necessary to flush
the fields that makes use of to ensure the correct entry is generated.

The domain for the query is inserted into the query using the output of
the method '_get_options_domain'. Part of the domain is the
'parent_state', which was updated in v16 (from 'account_move.state').

This commit appends the 'parent_state' to the list of fields flushed on
the account.move.line model before the closing entry query is executed.

closes odoo/enterprise#37133

X-original-commit: 6aa9e4ea369260926a73aa527061c7a374da0ed6
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Daniel Kosky (dako) <dako@odoo.com>