[FIX] approvals: custom view for `approval.product.line` in studio
Current behaviour:
When creating a custom menu that wants to display the list view of
the `approval.product.line`, we get a stack trace about the lack of
a "parent" field.

Expected behaviour:
Should not have a stacktrace when creating a custom list view of
`approval.product.line`.

Steps to reproduce:
- Install Approvals and Studio
- Create a custom menu with Studio based on the existing model
  `approval.product.line`
- Open the new menu -> stacktrace

Reason for the problem:
By default, the view for `approval.product.line` is a view that is
embedded by default, therefor in the domains for the attribute we
use the "parent" field to reference to the "parent" view js-side.
But when creating a menu via Studio, this assertion is not correct
anymore, since the view is not embedded.

Fix:
Create a separate list view for the model with a higher priority, so
that studio takes it when creating the view. Also remove the
attributes, since the fields don't exist on the model and we can't
find their value from the "parent", since it doesn't exist.

Affected versions:
- 14.0
- 15.0
- saas-15.2
- 16.0
- saas-16.1
- master

opw-3164422

closes odoo/enterprise#37179

X-original-commit: cff4423ef623e1483a67a6b4ed98fe5e081f89fa
Signed-off-by: Kevin Baptiste <kba@odoo.com>