[FIX] web_studio: edit selection fields values after closing studio
Steps to reproduce
==================

- Go to the contact form
- Open studio
- Add a selection field
- Close studio
- Open studio again
- Select the newly created selection field

The "Edit Values" link should be in the sidebar

Cause of the issue
==================

The link for editing the selection values is only displayed if the field
is a manual field.

https://github.com/odoo/enterprise/blob/ff6417e5df622c8f8a1c12884db616f70cc526a3/web_studio/static/src/legacy/xml/view_editor_sidebar.xml#L733

But if the selection has not been created in the same studio lifecycle,
`manual` is undefined. This is because, only some keys are loaded when
calling `get_views`

opw-3167203

closes odoo/enterprise#37036

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>