[FIX] snailmail_account_followup: cannot adapt type 'res.company'
If applied, this commit will solve the issue of creating a letter in snailmail.

Error: 'cannot adapt type res.company' while evaluating _cron_execute_followup().
This error comes when we create the letter in snailmail, the 'company_id' is
passed as an object like 'to_send_partner.company_id' but it will take the 'id'
of the record. so it should be 'to_send_partner.company_id.id'.

sentry - 3833502258

closes odoo/enterprise#38937

X-original-commit: 3421109191b6a6f0e7032127669b97567b537601
Signed-off-by: Quentin De Paoli <qdp@odoo.com>
Signed-off-by: William André (wan) <wan@odoo.com>