[IMP] project_enterprise: make task list view consistant
Purpose of this commit have consistent list views between
project.task and the sub-tasks / blocked by o2m so that when
a change is done in one place it also impacts the other
places.

So, in this commit add/update sub-task and blocked by page list
views to make them identical.

task-3165202

closes odoo/enterprise#37055

Related: odoo/odoo#112620
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>