[FIX] knowledge: properly display properties panel
Impacted versions:
16.0+

How to reproduce:
* open Knowledge and create an article
* create a child article of that article
* add a property for the child article
* create a root article (with no parent)
* move the root article under the first article (will become its child)

Current behavior:
* the property panel does not open unless forced reload

Expected behavior:
* the property panel opens with the same property as the other child article

Fix:
Handle the display of the property panel in useEffect instead of only onMounted.
The issue is that switching from an article to another will not re-trigger the
onMounted callback of the knowledge renderer, meaning that the condition to
display the property panel was not properly recomputed.

Task-3122001
Task-3186465

closes odoo/enterprise#37616

X-original-commit: 5a61060c73d56bd1eb7bb8b6cb53eb6c0f62492b
Signed-off-by: David Beguin (dbe) <dbe@odoo.com>