[FIX] account_accountant: bank statement balances not updated
When creating a bank statement from the bank recon list view or bank statement line form view, the ending balance is not properly computed in the default_get method.
This PR enables the computation of the ending balances, and displays the warning message on the statement form view.
See the corresponding community PR.

closes odoo/enterprise#37074

Related: odoo/odoo#112710
Signed-off-by: Laurent Smet <las@odoo.com>