[FIX] account_accountant: bank statement balances not updated
When creating a bank statement from the bank recon list view or bank statement line form view, the ending balance is not properly computed in the default_get method.
This PR enables the computation of the ending balances, and displays the warning message on the statement form view.
See the corresponding community PR.

closes odoo/enterprise#38131

X-original-commit: 2ce1651f86f626c6021a0858f7864c8f0377ec99
Related: odoo/odoo#115174
Signed-off-by: Laurent Smet <las@odoo.com>
Signed-off-by: Ayob Habib (ayh) <ayh@odoo.com>