[FIX] spreadsheet_dashboard_account_accountant: upgrade to latest version
This dashboard was generated pre-16.0 when the json format version for odoo
related stuff (odooVersion) was 4. The most up-to-date version in 16.0 is
odooVersion=5.

The next commit fixes the Accounting dashboard and was build with the current
version (5).
To reduce the noise in the next commit, I first upgrade the json file
to version 5.

Part-of: odoo/enterprise#37092