[REF] *: simplify concrete fields API - remove update prop
This commit, is part of a series of commits that aim to simplifie the
concrete fields API.

In this commit we will remove update prop from concrete fields. Now each
field will directly use this.props.record.update to make changes, and
handle the save in fields that need to (e.g. priority). As a consequence
of this, the record props need to be mandatory.

task-id 3179751

closes odoo/enterprise#37118

Related: odoo/odoo#112792
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>