[IMP] account_avatax: autocomplete latitude and longitude
Before this PR when validating an address with Avalara the latitude and longitude of the response given by their API wasn't used. Since we also have a module named "Geo localization" that provide a way to get the geo localization of a user, we decided that it would be nice that when validating the address the geo localization field would change at the same time.

closes odoo/enterprise#37157

Task-id: 3186071
Related: odoo/odoo#113210
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>