[FIX] sale_subscription: prevent compare date with bool
Steps to reproduce:
- install sale_subscription and stock modules;
- create two products, one recurring and one storable with minimum one unit on hand;
- create a subscription with these two products (without start date).

Issue:
A traceback appears.

Cause:
In the case where we don't give a start date on the quotation,
we will use functions of other modules (due to the storable product) which will call
the function `_compute_last_invoice_date` before the start date has been set
(which triggers the error).
In the flow, this date is set via the `_confirm_subscription` function afterwards.

Solution:
Check that the date exists before using it in the comparison.

opw-3179800

closes odoo/enterprise#37178

X-original-commit: c33f133afb93188822e334d0a80f0eb4daf101a9
Signed-off-by: Arnaud Joset <arj@odoo.com>
Signed-off-by: Lefebvre Thomas (thle) <thle@odoo.com>